Разбивает код на модули с точкой входа в main.js
