+++ js/utile.js
@@ -1,0 +1,30 @@
+const getRandomPositiveInteger = (n1, n2) => {
+  if (n1 >= 0 && n2 >= 0 && n1 !== n2) {
+    const lower = Math.ceil(Math.min(n1, n2));
+    const upper = Math.floor(Math.max(n1, n2));
+
+    return Math.floor(Math.random() * (upper - lower + 1)) + lower;
+  }
+
+  return NaN;
+};
+
+const getRandomPositiveFloat = (n1, n2, digits) => {
+  if (n1 >= 0 && n2 >= 0 && n1 !== n2 && digits > 0) {
+    const lower = Math.min(n1, n2);
+    const upper = Math.max(n1, n2);
+    const result = Math.random() * (upper - lower) + lower;
+
+    return Number(result.toFixed(digits));
+  }
+
+  return NaN;
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {
+  getRandomPositiveInteger,
+  getRandomPositiveFloat,
+  getRandomArrayElement,
+};
